refactor update 3/2

+++ layouts.sh
@@ -7,16 +7,14 @@ ARGS="$@"
 
 usage() {
     printf '%s\n' "Usage: $(basename $0) <save|load|open|replace|list|delete> <layout>"
-    test -z $1 && exit 0 || exit $1
+    test -z $1 || exit $1
 }
 
 # save all current visible windows to a layout file
 saveLayout() {
     LAY=$1
     # clean file for new run
-    if [ -e $LAYOUTDIR/layout.$LAY ]; then
-        rm $LAYOUTDIR/layout.$LAY 2> /dev/null
-    fi
+    test -e $LAYOUTDIR/layout.$LAY && rm $LAYOUTDIR/layout.$LAY 2> /dev/null
 
     # sort based on x values
     windowLoop=$(lsw | wc -l)
@@ -28,7 +26,7 @@ saveLayout() {
         psid=$(wclass.sh p $wid)
         window=$(wclass.sh c $wid)
 
-        # sadly we have to clean the window classes to their called names
+        # rename the window classes to their called arguments at their launch
         test "$window" = "URxvt" && \
             window=$(ps $psid | tail -1 | perl -p -e 's/^.*?urxvt/urxvt/')
 
@@ -46,10 +44,10 @@ saveLayout() {
 # try and position the currently open windows to a layout
 loadLayout() {
     LAY=$1
-    if [ ! -e $LAYOUTDIR/layout.$LAY ]; then
+    test ! -e $LAYOUTDIR/layout.$LAY && {
         printf '%s\n' "Layout does not exist, exiting ..." >&2
         exit 1
-    fi
+    }    
 
     windowLoop=$(cat $LAYOUTDIR/layout.$LAY | wc -l)
     lswSort=$(lsw | xargs wattr xi | sort -n | sed "s/^[0-9]* //")
@@ -82,19 +80,20 @@ loadLayout() {
 # load the entire layout regardless of currently open windows
 openLayout() {
     LAY=$1
-    if [ ! -e $LAYOUTDIR/layout.$LAY ]; then
+    test ! -e $LAYOUTDIR/layout.$LAY && {
         printf '%s\n' "Layout does not exist, exiting ..." >&2
         exit 1
-    fi
+    }    
 
     windowLoop=$(cat $LAYOUTDIR/layout.$LAY | wc -l)
     for i in $(seq $windowLoop); do
         lswOpen=$(lsw | wc -w)
 
+        # load program
         program=$(sed "$i!d" < $LAYOUTDIR/layout.$LAY | cut -d\  -f 5-)
         $program &
 
-        # wait for window to appear
+        # wait for window to appear - this fking sucks
         while [ $lswOpen -eq $(lsw | wc -w) ]; do
             sleep 0.5
         done
@@ -112,37 +111,35 @@ openLayout() {
 # but also open windows when a wid cannot be found
 replaceLayout() {
     LAY=$1
-    if [ ! -e $LAYOUTDIR/layout.$LAY ]; then
+    test ! -e $LAYOUTDIR/layout.$LAY && {
         printf '%s\n' "Layout does not exist, exiting ..." >&2
         exit 1
-    fi
+    }    
 
 }
 
 listLayout() {
     LAY=$1
-    if [ ! -e $LAYOUTDIR/layout.$LAY ]; then
+    test ! -e $LAYOUTDIR/layout.$LAY && {
         printf '%s\n' "Layout does not exist, exiting ..." >&2
         exit 1
-    fi
+    }    
 
     cat $LAYOUTDIR/layout.$LAY
 }
 
 deleteLayout() {
     LAY=$1
-    if [ ! -e $LAYOUTDIR/layout.$LAY ]; then
+    test ! -e $LAYOUTDIR/layout.$LAY && {
         printf '%s\n' "Layout does not exist, exiting ..." >&2
         exit 1
-    else
+    } || \    
         rm $LAYOUTDIR/layout.$LAY 2> /dev/null
-    fi
 }
 
 main() {
     . fyrerc.sh
 
-    # test arguments
     test -z $2 && usage 1
     case $1 in
         s|save)    saveLayout $2    ;;
@@ -151,8 +148,7 @@ main() {
         d|delete)  deleteLayout $2  ;;
         r|replace) replaceLayout $2 ;;
         ls|list)   listLayout $2    ;;
-        h|help)    usage            ;;
-        *)         usage            ;;
+        *)         usage 0          ;;
     esac
 }
